Fix rendering/plotting of arcs with tiny angle and huge radius.
Fixes https://gitlab.com/kicad/code/kicad/-/issues/17343

(cherry picked from commit 4d66a8ebdbdba711e34dd6530606b2e3a7bf3c10)

+++ include/plotters/plotter_gerber.h
@@ -292,7 +292,7 @@ protected:
      * be initialized before calling it if needed.
      */
     void plotArc( const VECTOR2I& aCenter, const EDA_ANGLE& aStartAngle, const EDA_ANGLE& aEndAngle,
-                  int aRadius, bool aPlotInRegion );
+                  double aRadius, bool aPlotInRegion );
     void plotArc( const SHAPE_ARC& aArc, bool aPlotInRegion );
 
     /**
